requirements.txt
mysql-connector-python
nltk
tweepy
Flask
Flask-Cors